requirements.txt
blinker==1.9.0
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
cryptography==45.0.6
Flask==3.1.1
flask-cors==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.1
lxml==6.0.0
MarkupSafe==3.0.2
nltk==3.9.1
numpy==2.3.2
pandas==2.3.1
pdfminer.six==20250506
pycparser==2.22
python-dateutil==2.9.0.post0
python-docx==1.2.0
pytz==2025.2
regex==2025.7.34
scikit-learn==1.7.1
scipy==1.16.1
six==1.17.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
Werkzeug==3.1.3
